Smarter setup

+++ install.sh
@@ -36,6 +36,7 @@ install_from_host_home() {
     }
     for dir_name in bin .local/bin/cdpp .local/bin/localhist my-home .ssh .vim bb-cert; do
         (
+            [[ -e ${host_home}/${dir_name} ]] || exit 0
             [[ -d $dir_name ]] || {
                 mkdir -p $dir_name || die "Can't create ~/${dir_name}"
             }
@@ -44,9 +45,12 @@ install_from_host_home() {
         )
     done
     for file_name in .cdpprc .localhistrc; do
+        [[ -e ${host_home}/${file_name} ]] || continue
         cp ${host_home}/${file_name} ~/${file_name}
     done
-    echo "source ~/bin/bashrc-common # Added by $scriptId" >> ~/.bashrc
+    grep -Eq "^source ~/bashrc-common" ~/.bashrc || {
+        echo "source ~/bin/bashrc-common # Added by $scriptId" >> ~/.bashrc
+    }
     ln -sf bin/inputrc .inputrc
     ln -sf my-home/gitconfig .gitconfig
     [[ -e ~/projects ]] || mkdir ~/projects
